tf_agents/pyproject.toml
[tool.poetry]
name = "cogment-verse-tf-agents"
version = "0.1.0"
description = ""
authors = ["AI Redefined Inc. <dev+cogment@ai-r.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
cogment-verse = {path = "../base_python", develop = true}
tensorflow = {version="2.6.2", optional = true}
torch = {version="1.8.1", optional = true}
tensorflow-probability = "^0.14.1"
opencv-python-headless = "4.4.0.40"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
taskipy = "^1.9.0"

[tool.poetry.extras]
with_tf = ["tensorflow", "torch"]

[tool.taskipy.tasks]
generate_protos = "./scripts/generate_protos.sh"
build = "task generate_protos"
start="python -m main"

[tool.black]
line-length = 120
target-version = ["py37"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
